decoder.quirc: implement center point computation.
This is done in cython because quirc does not appear to provide this
functionality. Uses the same algorithm that I added to qrdec, although
this computation is not done at higher precision so the results may
not match exactly (the bounding boxes that quirc produces are slightly
different than the ones produced by qrdec anyway).