ci: retool for the last time, I swear.
This switches to using the manylinux_2_24 image for the linux builds,
with dependencies preinstalled.

Poetry is allowed to make its own venv in the non-docker builds as
there's no reason to prevent it on the builds that aren't running in
images with preinstalled dependencies. It turns out poetry's need
to aggressively uninstall dependencies that aren't in its lockfile
actually interferes with some of the CI environments, so don't do it.

This time publishing is definitely fixed, for real.