Add image loader.
This is written in such a way that it (ab)uses python's dynamic object
reflection to check the image container type without needing to
actually import any dependencies. The input format support has been
expanded to include cv2.UMat (which gets converted to numpy.ndarray),
numpy.ndarray, PIL.Image (theoretically, this is currently untested)
for images loaded with Pillow, and data that is already in a byte
buffer with a width and height.