docs: start documenting things.
This is an experiment. The published documentation is stored in a
separate repository. I want to be able to store binary data in the
documentation (e.g. pictures) without junking up the source
repository. This makes the ergonomics of publishing the documentation
a bit clunkier, but CI automation means it doesn't actually matter.

Docs are only generated for master and tags so that poorly named
branches can't clobber tag docs.