docs: add a hack for absolute image URI production.
Sphinx rewrites all image URIs to be relative to the _static directory
which is bundled with the build. However, I absolutely do not want to
store images in this repository, and pulling them in during the build
process would just result in them being copied into the output
directory. Since the goal is to have many copies of the documentation
stored int the same repository (one for each version plus master
branch), I don't want each published version to

This could theoretically be solved using symlinks to a shared image
directory from each published version, but that complicates the
publishing process. I may try to take that approach later, but at some
point this simply became a challenge to see if I could make this
approach work. It's very convenient that the Sphinx configuration is
just an evaluated python script that allows this sort of degenerate
monkey patching to occur.

This is tucked behind a new image directive, gimage, so as to be
opt-in. Since using this expects the documentation to be compiled
without actual access to the images, it will only be able to produce
images for a very specific html layout. Other output formats, such as
PDF, won't build properly, but I don't care much about that.