decoders.quirc: handle decoding errors correctly.
If something that looks like a QR code is detected but it cannot
actually be decoded, then quirc_decode returns an error code. In that
case, data.data_type and data.payload contain meaningless values.

Theoretically, we could still produce the bounding box, but that adds
weird edge cases to the client handling (it would require the user to
check each returned QRCode to make sure it had been decoded
correctly), which seems like a suboptimal API. This may be worth
revisiting in the future, however.