decoders: use sane data conversion defaults.
In the general case, returning data as bytes isn't particularly
helpful. While it's true that in some advanced cases, converting data
requires additional information that the library doesn't have access
to, I think it's better for the basic case to work out of the box
while providing enough customization capability for more advanced use
cases to still be practical.

This change is straining against the load-bearing copy-paste technique,
which I don't like, but I'll save an attempt to extract the common
functionality until later.