docs.theme: configure things that weren't meant to be configured.
I thought about it and realized, while I'm on the monkey patching hell
train, I can abuse the theme implementation to get it to do some stuff
I'd like, but that isn't exposed as configurable for one reason or
another.

This is: making the left sidebar the only persistent TOC that includes
subsections as well as pages, rather than having the subsections for
the current page show up on the right. Also, while the right-side
section listing can be disabled without monkey patching, it can only
be done on a per-page basis as far as I can tell, which is cumbersome
at best.