CI: improvements to doc build.
Store different symlinks for the release/dev documentation. `latest`
will always point to the latest release version and `latest-dev`
points to the latest version generated from the master branch.

Since this workflow mucks around with the deqr-doc git repository, also
change the concurrency setting so that it will not ever try to run two
doc gen workflows in parallel.