CI: retool and add python 3.11
This was (ab)using poetry 1.1.x by installing it in the same virtual
environment as the project dependencies. This is a bad idea for a
variety of reasons, and it works even less well with newer versions of
poetry. Unfortunately, the most straightforward option, letting poetry
manage the virtualenv, doesn't work well with the manylinux image
containing multiple python versions. This could be worked around, but
it doesn't seem worth it. Poetry is contained to its own virtualenv,
which seems to work okay in development with multiple python versions
as long as it is being run in a pre-created virtual environment. The
PyPA `build` tool seems to be the easiest way to build for a specific
version, as `poetry build` produces wrong-named outputs (probably due
to bugs).

poetry-dynamic-versioning is a nice concept but it has some really
annoying downsides, the most obvious of which is that it will cause a
variety of poetry commands to fail if run outside of a git repository
with just the message `This does not appear to be a Git project`. I
expect this could be fixed with little effort (and is maybe even a
regression), but the level of automagic is slightly concerning.