add a changelog
it doesn't look like Poetry has a field specifier for a changelog
(which probably means pypi doesn't either). However, it's nice to have
this information readily available on the pypi page, in my opinion.

Also, while I like putting the changelog into the tag, it's also good
to have it here. Plus, this will make generating the tag changelog
easier.