CI: try to be really aggressive about eliminating build artifacts
For some reason the Linux wheels published by the CI job all had the
compiled cp38 objects in them, which broke the wheels for 3.9, 3.10,
and 3.11. I could not reproduce this manually stepping through the CI
job in a local version of the container, but it wasn't a perfect
replication (I cloned the source directly rather than using a bind
mount, which github actions does). Anyway, in theory, it should be
impossible for wheels to pick up cached/wrong object files if they
don't exist.