build: attempt to eliminate debug information
I don't know if this switch works on any compiler except clang on
macOS, but -Wl,-s does not work at all on macOS. This prevents
references to the build machine's filesystem from being embedded in
the resulting compiled objects (as long as the Python installation
also doesn't cause the linker to embed spurious rpath headers
in _sysconfigdata_[platform].py). Perhaps -Wl,-s should be added as a
linker argument if -g0 doesn't work on linux/gcc.

I'm also really not sure if the cython-produced dylibs should link to
libpython statically or dynamically. It appears that the build process
prefers static linking, and I'm not particularly inclined to tinker
with the defaults.