remove support for python 3.8 and 3.9 and add 3.13 builds
3.8 is EOL and 3.9 is EOL in about 2 months.

3.13t required a bit of reworking of some of the CI scripts as one of
poetry's dependencies, zstandard, fails to compile. Fortunately,
poetry_core, which is what is actually needed for building works fine
with 3.13t. This did necessitate reworking the way the version
checking was done, but now it is even better because it also checks
that the tag version matches the project version on the publishing
workflow.

Also some of the publishing jobs had to be switched to using twine,
though poetry can still be used in the manylinux image since it runs
under 3.13 regular there.