CI: generate build container makefile
The old version was an interesting hack but utterly unmaintainable, as
it fell apart upon encountering a light breeze and was completely
impossible to debug. This also renames the dockerfile in preparation
for adding support for the musllinux image.

Should the generator have been written in python? That would certainly
have made more sense. However, I'm not here to make sense.